Move PolarisMetaStoreManagerImpl into the "transactional" package per PR suggestion
